+++ src-ts/tools/learn/course-details/course-curriculum/tc-academy-policy-modal/TcAcademyPolicyModal.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../lib/styles/includes';
+@import '../../../../../lib/styles/includes';
 
 .container {
