+++ src-ts/tools/learn/course-details/course-curriculum/tc-academy-policy-modal/index.ts
